Deep-Reinforcement-Stock-Trading/requirements.txt
tensorflow==2.20.0
numpy==2.3.2
scipy==1.16.1
pandas==2.3.1
matplotlib==3.10.5
statsmodels==0.14.5
h5py==3.14.0
yfinance==0.2.52
python-dotenv==1.0.1
requests==2.32.4
